# api/openapi.yaml
openapi: 3.0.3
info:
  title: Insurance Claims API
  version: 2.6.0
  description: |
    Contract of the claims settlement platform: policies, claims, supporting
    documents, expertise, payouts, customers and tariffs.

    The document is maintained by hand and is the contract of record for the
    `claims-api-quarkus` service. Neither smallrye-openapi nor any other runtime
    generator is on the classpath, so the file is never overwritten at build time.
  contact:
    name: Claims Platform Team
    email: claims-platform@insurer.example.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api
    description: Local developer instance; the /api prefix comes from @ApplicationPath("/api")
  - url: https://claims.insurer.example.com/api
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /v1/claims:
    get:
      operationId: searchClaims
      summary: Search claims
      description: Filters are bound as a single bean parameter on the server side.
      tags: [claims]
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/TenantHeaderParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ClaimStatus'
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ClaimType'
        - name: policyId
          in: query
          required: false
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: Page of claims
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: registerClaim
      summary: Register a new claim
      tags: [claims]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCreateRequest'
      responses:
        '201':
          description: Claim registered
          headers:
            Location:
              description: URI of the registered claim
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Claim'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/claims/{id}:
    parameters:
      - $ref: '#/components/parameters/ClaimIdParam'
    get:
      operationId: getClaim
      summary: Get a single claim
      tags: [claims]
      responses:
        '200':
          description: Claim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Claim'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: patchClaim
      summary: Update the status, description or approved amount of a claim
      tags: [claims]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimPatchRequest'
      responses:
        '200':
          description: Claim updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Claim'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/customers:
    get:
      operationId: listCustomers
      summary: List customers
      tags: [customers]
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: lastName
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
      responses:
        '200':
          description: Page of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createCustomer
      summary: Register a customer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateRequest'
      responses:
        '201':
          description: Customer created
          headers:
            Location:
              description: URI of the created customer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/customers/{id}:
    parameters:
      - $ref: '#/components/parameters/CustomerIdParam'
    get:
      operationId: getCustomer
      summary: Get a single customer
      tags: [customers]
      responses:
        '200':
          description: Customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replaceCustomer
      summary: Replace the personal data of a customer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/expertises:
    get:
      operationId: listExpertises
      summary: List expertise records
      tags: [expertise]
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: claimId
          in: query
          required: false
          schema:
            type: string
            maxLength: 40
        - name: verdict
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/VerdictType'
      responses:
        '200':
          description: Page of expertise records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertisePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createExpertise
      summary: Order an expertise for a claim
      tags: [expertise]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpertiseCreateRequest'
      responses:
        '201':
          description: Expertise ordered
          headers:
            Location:
              description: URI of the created expertise record
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expertise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/expertises/{id}:
    get:
      operationId: getExpertise
      summary: Get a single expertise record
      description: Returns 206 while the expert still needs additional data.
      tags: [expertise]
      parameters:
        - $ref: '#/components/parameters/ExpertiseIdParam'
      responses:
        '200':
          description: Expertise with a final verdict
          headers:
            X-Expertise-Verdict:
              description: Verdict duplicated as a header
              schema:
                $ref: '#/components/schemas/VerdictType'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expertise'
        '206':
          description: Expertise is incomplete and awaits additional data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expertise'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/expertises/{id}/verdict:
    patch:
      operationId: applyExpertiseVerdict
      summary: Record the verdict of an expertise
      tags: [expertise]
      parameters:
        - $ref: '#/components/parameters/ExpertiseIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerdictRequest'
      responses:
        '200':
          description: Verdict recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expertise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/expertises/{id}/report:
    parameters:
      - $ref: '#/components/parameters/ExpertiseIdParam'
    post:
      operationId: uploadExpertiseReport
      summary: Upload the scanned expertise report
      tags: [expertise]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Scanned report
      responses:
        '200':
          description: Report attached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expertise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: downloadExpertiseReport
      summary: Download the rendered expertise report
      tags: [expertise]
      responses:
        '200':
          description: Report payload
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v0/claims/{id}:
    get:
      operationId: getLegacyClaim
      summary: Frozen v0 claim lookup kept for the legacy back office
      description: Deprecated. Use the v1 claim endpoints instead.
      deprecated: true
      tags: [legacy]
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ClaimIdParam'
      responses:
        '200':
          description: Claim in the v1 representation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Claim'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Operator access token issued by the corporate identity provider
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Static key used by the legacy back office
  parameters:
    OffsetParam:
      name: offset
      in: query
      required: false
      description: Index of the first returned element
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    LimitParam:
      name: limit
      in: query
      required: false
      description: Maximum number of returned elements
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 25
    TenantHeaderParam:
      name: X-Tenant-Id
      in: header
      required: false
      description: Tenant selector, bound through the bean parameter of the search endpoint
      schema:
        type: string
        maxLength: 40
        default: default
...
    ClaimIdParam:
      name: id
      in: path
      required: true
      description: Claim identifier
      schema:
        type: string
        maxLength: 40
...
    CustomerIdParam:
      name: id
      in: path
      required: true
      description: Customer identifier
      schema:
        type: string
        maxLength: 40
...
    ExpertiseIdParam:
      name: id
      in: path
      required: true
      description: Expertise identifier
      schema:
        type: string
        maxLength: 40
  responses:
    BadRequest:
      description: The request payload or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller is authenticated but not allowed to perform the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The operation conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
        resource:
          type: string
        timestamp:
          type: string
          format: date-time
        violations:
          type: array
          items:
            type: string
    PageMeta:
      type: object
      required: [offset, limit, total]
      properties:
        offset:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    MoneyAmount:
      type: object
      required: [value, currency]
      properties:
        value:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: RUB
...
    ClaimType:
      type: string
      enum: [PROPERTY_DAMAGE, THEFT, FLOOD, FIRE, MEDICAL, LIABILITY]
    ClaimStatus:
      type: string
      enum: [REGISTERED, UNDER_REVIEW, EXPERTISE, APPROVED, REJECTED, PAID, CLOSED]
    Claim:
      type: object
      required: [id, claimNumber, policyId, customerId, type, status, incidentDate, claimedAmount]
      properties:
        id:
          type: string
        claimNumber:
          type: string
          pattern: '^CLM-[0-9]{8}$'
        policyId:
          type: string
        customerId:
          type: string
        type:
          $ref: '#/components/schemas/ClaimType'
        status:
          $ref: '#/components/schemas/ClaimStatus'
        incidentDate:
          type: string
          format: date
        description:
          type: string
          maxLength: 4000
        claimedAmount:
          $ref: '#/components/schemas/MoneyAmount'
        approvedAmount:
          $ref: '#/components/schemas/MoneyAmount'
        registeredAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ClaimCreateRequest:
      type: object
      required: [policyId, customerId, type, incidentDate, description, claimedAmount]
      properties:
        policyId:
          type: string
          maxLength: 40
        customerId:
          type: string
          maxLength: 40
        type:
          $ref: '#/components/schemas/ClaimType'
        incidentDate:
          type: string
          format: date
        description:
          type: string
          minLength: 20
          maxLength: 4000
        claimedAmount:
          $ref: '#/components/schemas/MoneyAmount'
    ClaimPatchRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ClaimStatus'
        description:
          type: string
          maxLength: 4000
        approvedAmount:
          $ref: '#/components/schemas/MoneyAmount'
    ClaimPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Claim'
...
    Contact:
      type: object
      required: [email, phone]
      properties:
        email:
          type: string
          format: email
          maxLength: 160
        phone:
          type: string
          pattern: '^\+[0-9]{10,15}$'
        preferredChannel:
          type: string
          maxLength: 200
    Customer:
      type: object
      required: [id, firstName, lastName, birthDate, contact]
      properties:
        id:
          type: string
        firstName:
          type: string
          maxLength: 80
        lastName:
          type: string
          maxLength: 80
        birthDate:
          type: string
          format: date
        contact:
          $ref: '#/components/schemas/Contact'
        taxNumber:
          type: string
          pattern: '^[0-9]{12}$'
        registeredAt:
          type: string
          format: date-time
    CustomerCreateRequest:
      type: object
      required: [firstName, lastName, birthDate, contact]
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 80
        lastName:
          type: string
          minLength: 2
          maxLength: 80
        birthDate:
          type: string
          format: date
        contact:
          $ref: '#/components/schemas/Contact'
        taxNumber:
          type: string
          pattern: '^[0-9]{12}$'
    CustomerUpdateRequest:
      type: object
      required: [firstName, lastName, contact]
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 80
        lastName:
          type: string
          minLength: 2
          maxLength: 80
        contact:
          $ref: '#/components/schemas/Contact'
    CustomerPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
...
    VerdictType:
      type: string
      enum: [CONFIRMED, PARTIALLY_CONFIRMED, REJECTED, NEEDS_MORE_DATA]
    Expertise:
      type: object
      required: [id, claimId, expertName, organisation, inspectionDate, verdict]
      properties:
        id:
          type: string
        claimId:
          type: string
        expertName:
          type: string
          maxLength: 160
        organisation:
          type: string
          maxLength: 160
        inspectionDate:
          type: string
          format: date
        verdict:
          $ref: '#/components/schemas/VerdictType'
        assessedAmount:
          $ref: '#/components/schemas/MoneyAmount'
        reportFileName:
          type: string
        createdAt:
          type: string
          format: date-time
    ExpertiseCreateRequest:
      type: object
      required: [claimId, expertName, organisation, inspectionDate]
      properties:
        claimId:
          type: string
          maxLength: 40
        expertName:
          type: string
          maxLength: 160
        organisation:
          type: string
          maxLength: 160
        inspectionDate:
          type: string
          format: date
    VerdictRequest:
      type: object
      required: [verdict, assessedAmount]
      properties:
        verdict:
          $ref: '#/components/schemas/VerdictType'
        assessedAmount:
          $ref: '#/components/schemas/MoneyAmount'
        rationale:
          type: string
          maxLength: 2000
    ExpertisePage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Expertise'